[RUSAA-69] refactor(ingest-consumer): split into sub-modules under 600-line cap
- Split ingest_consumer.rs (647 lines) into three focused modules:
  - mod.rs (369 lines): spawn function + all tests
  - db.rs (192 lines): DB update functions + TOTAL_PIPELINE_STAGES
  - sse.rs (103 lines): SSE formatting + OTel tracing helpers
- Fix PR body: add Closes #166, remove bare tracker identifiers
  from unrelated issues, remove Paperclip line
- All 159 tests pass, clippy clean
